+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Add getters/setters for `JsonPropertyFilter` class to obtain/set the exclude/include filters.
+- Add the posibility to define the symbols of include/exclude.
 
 ### Changed
 - All constructor arguments of the `JsonPropertyFilter` class are optionals.